requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
mlxtend==0.23.0
pandas
numpy
google-generativeai
requests
pillow
gunicorn
